+++ packages/target-tauri/src-tauri/src/html_window/mod.rs
@@ -67,6 +67,7 @@ pub(crate) async fn open_html_window(
             // window already exists focus it - android and iOS don't have have the function
             // and those platforms also don't have multiple windows
             window.show()?;
+            window.set_focus()?;
             return Ok(());
         }
     }